requirements.txt
gradio==4.44.1
gradio_client==1.3.0
fastapi==0.112.4
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
tavily-python>=0.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
accelerate
SpeechRecognition
diffusers
torch
pillow
transformers
safetensors
huggingface_hub==0.36.0
sentencepiece
